requirements.txt
torch>=1.8.1
torchvision>=0.9.1
Pillow